Organizing Notebook with Markdown, Serializing Data, and Building Model
I added more markdown in the jupyter notebook to better organize the
sections. I now serialize the data once loaded (in /data folder) for
ease of use + data only has to be parsed one time now, and can just be
resused. And now have well-defined functions for building the model:
that is the generator, the discriminator, and their parts in combination
(the GAN).